--- app/components/animation-lottie.jsx
@@ -1,20 +1,0 @@
-"use client";
-
-import Lottie from "lottie-react";
-
-const AnimationLottie = ({ animationPath, width }) => {
-  const defaultOptions = {
-    loop: true,
-    autoplay: true,
-    animationData: animationPath,
-    style: {
-      width: '95%',
-    }
-  };
-
-  return (
-    <Lottie {...defaultOptions} />
-  );
-};
-
-export default AnimationLottie;
