+++ app/components/helper/animation-lottie.jsx
@@ -1,0 +1,20 @@
+"use client";
+
+import Lottie from "lottie-react";
+
+const AnimationLottie = ({ animationPath, width }) => {
+  const defaultOptions = {
+    loop: true,
+    autoplay: true,
+    animationData: animationPath,
+    style: {
+      width: '95%',
+    }
+  };
+
+  return (
+    <Lottie {...defaultOptions} />
+  );
+};
+
+export default AnimationLottie;
