+++ app/components/homepage/education.jsx
@@ -3,8 +3,8 @@ import Image from "next/image";
 
 import { educations } from "@/utils/data/educations";
 import { BsPersonWorkspace } from "react-icons/bs";
-import AnimationLottie from "../animation-lottie";
-import GlowCard from "../glow-card";
+import AnimationLottie from "../helper/animation-lottie";
+import GlowCard from "../helper/glow-card";
 import lottieFile from '/public/lottie/study.json';
 
 function Education() {
