+++ app/components/homepage/experience.jsx
@@ -3,8 +3,8 @@
 import { experiences } from "@/utils/data/experience";
 import Image from "next/image";
 import { BsPersonWorkspace } from "react-icons/bs";
-import AnimationLottie from "../animation-lottie";
-import GlowCard from "../glow-card";
+import AnimationLottie from "../helper/animation-lottie";
+import GlowCard from "../helper/glow-card";
 import experience from '/public/lottie/code.json';
 
 function Experience() {
